edit and move types
